feat: login 시 myInfo 저장
