+++ src/components/utilComponents/AutoLoginPresence.tsx
@@ -1,5 +1,5 @@
 import { useAutoLoginMutation } from '@src/hooks/apis/user';
-import { accessTokenState } from '@src/recoil';
+import { accessTokenState } from '@src/recoil/permissionAtoms';
 import { ReactNode, useEffect } from 'react';
 import { useRecoilValue } from 'recoil';
