+++ src/hooks/apis/user/useAutoLoginMutation.ts
@@ -6,7 +6,7 @@ import { AutoLoginResponse } from '@src/types/loginTypes';
 import { usePopup } from '@src/hooks';
 import { PATH } from '@src/constants/urls';
 import { AUTO_LOGIN_API } from '@src/constants/apiUrl';
-import { accessTokenState } from '@src/recoil';
+import { accessTokenState } from '@src/recoil/permissionAtoms';
 import { ERROR_MESSAGE } from '@src/constants/message';
 import { POPUP_DURATION } from '@src/constants/duration';
