+++ src/hooks/apis/user/useLogoutMutation.ts
@@ -1,14 +1,15 @@
 import axios from 'axios';
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 import { useMutation } from '@tanstack/react-query';
 import { usePopup } from '@src/hooks';
 import { LOGOUT_API } from '@src/constants/apiUrl';
 import { POPUP_MESSAGE } from '@src/constants/message';
 import { POPUP_DURATION } from '@src/constants/duration';
-import { accessTokenState } from '@src/recoil';
+import { accessTokenState, myProfileState } from '@src/recoil/permissionAtoms';
 
 function useLogoutMutation() {
   const [accessToken, setAcessToken] = useRecoilState(accessTokenState);
+  const setMyProfile = useSetRecoilState(myProfileState);
   const { openAndDeletePopup } = usePopup();
 
   const mutationFn = () =>
@@ -17,6 +18,7 @@ function useLogoutMutation() {
     });
   const onSuccess = () => {
     setAcessToken(undefined);
+    setMyProfile(undefined);
     openAndDeletePopup({
       message: POPUP_MESSAGE.logoutSuccess,
       duration: POPUP_DURATION.medium,
