+++ src/types/loginTypes.ts
@@ -1,3 +1,5 @@
+import { MyProfile } from './userTypes';
+
 export interface LoginRequest {
   id: string;
   pw: string;
@@ -6,6 +8,7 @@ export interface LoginRequest {
 export interface LoginResponse {
   accessToken: string;
   refreshToken: string;
+  myInfo: MyProfile;
 }
 
 export interface AutoLoginResponse {
