+++ src/types/userTypes.ts
@@ -7,3 +7,14 @@ export interface Profile {
   role: string;
   homepage: string;
 }
+
+export interface MyProfile {
+  id: string;
+  role: string;
+  name: string;
+  nickname: string;
+  email: string;
+  introduce: string;
+  homepage: string;
+  image: string;
+}
